rename sign-up to register
