+++ src/app/app.routes.ts
@@ -3,7 +3,7 @@ import { Routes } from '@angular/router';
 export const routes: Routes = [
   {
     title: 'Increase financial knowledge and develop crypto currencies (sentiments) investment skills with THD(C)',
-    path: 'sign-up',
-    loadChildren: () => import('./modules/sign-up/sign-up.module').then(m => m.SignUpModule),
+    path: 'register',
+    loadChildren: () => import('./modules/register/register.module').then(m => m.RegisterModule),
   },
 ];
