+++ src/app/modules/register/components/register/register.config.ts
