+++ src/app/modules/register/components/register/register.model.ts
@@ -1,3 +1,3 @@
 import { User } from '../../../../shared/models/user.model';
 
-export type SignUpRequest = User;
+export type RegisterRequest = User;
