+++ src/app/modules/register/register.routes.ts
@@ -1,0 +1,9 @@
+import { Routes } from '@angular/router';
+import { RegisterComponent } from './components/register/register.component';
+
+export const routes: Routes = [
+  {
+    path: '**',
+    component: RegisterComponent,
+  },
+];
