--- src/app/modules/sign-up/sign-up.routes.ts
@@ -1,9 +1,0 @@
-import { Routes } from '@angular/router';
-import { SignUpComponent } from './components/sign-up/sign-up.component';
-
-export const routes: Routes = [
-  {
-    path: '**',
-    component: SignUpComponent,
-  },
-];
